+++ src/renderer/board/tools/lasso.tool.ts
@@ -3,6 +3,6 @@ import { SelectionTool } from '@/renderer/board/tools/selection.tool';
 
 export class LassoTool extends SelectionTool {
   constructor() {
-    super('lasso', 'Lasso', new LassoRegion());
+    super('lasso', 'Lasso', 'crosshair', new LassoRegion());
   }
 }